+++ nested_admin/__init__.py
@@ -152,10 +152,7 @@ def all_valid(original_all_valid, formsets):
                         # Force Django to try to save the instance,
                         # since we need it for the fk to work
                         changed_data = parent_form.fields.keys()
-                        if django.VERSION > (1, 9):
-                            parent_form.__dict__['changed_data'] = changed_data
-                        else:
-                            parent_form._changed_data = changed_data
+                        parent_form.__dict__['changed_data'] = changed_data
                 if not hasattr(parent_form, 'parent_formset'):
                     break
                 parent_form.parent_formset._errors = None