+++ nested_admin/formsets.py
@@ -202,10 +202,7 @@ class NestedInlineFormSetMixin(object):
                         form.data[form.add_prefix(sort_field)] = six.text_type(i)
 
                     # Force recalculation of changed_data
-                    if django.VERSION > (1, 9):
-                        form.__dict__.pop('changed_data', None)
-                    else:
-                        form._changed_data = None
+                    form.__dict__.pop('changed_data', None)
 
                 i += 1
 
@@ -266,12 +263,6 @@ class NestedInlineFormSetMixin(object):
             self.__queryset = qs
         return self.__queryset
 
-    if django.VERSION < (1, 9):
-        def delete_existing(self, obj, commit=True):
-            """Deletes an existing model instance."""
-            if commit:
-                obj.delete()
-
     def save_existing_objects(self, initial_forms=None, commit=True):
         """
         Identical to parent class, except ``self.initial_forms`` is replaced
@@ -310,10 +301,7 @@ class NestedInlineFormSetMixin(object):
                         form.data[form.add_prefix(pk_name)] = ''
 
                     if not form.has_changed():
-                        if django.VERSION > (1, 9):
-                            form.__dict__['changed_data'].append(pk_name)
-                        else:
-                            form._changed_data.append(pk_name)
+                        form.__dict__['changed_data'].append(pk_name)
 
                     saved_instances.extend(self.save_new_objects([form], commit))
                     continue