+++ nested_admin/templatetags/nested_admin.py
@@ -4,7 +4,6 @@ import json
 
 import django
 from django import template
-from django.apps import apps
 from django.conf import settings
 from django.contrib.admin.options import InlineModelAdmin
 from django.utils.safestring import mark_safe
@@ -13,26 +12,7 @@ from django.utils.html import escape
 register = template.Library()
 
 
-if django.VERSION >= (1, 10):
-    from django.templatetags.static import static as _static
-else:
-    _static = None
-
-
 django_version = django.VERSION[:2]
-
-
-@register.simple_tag
-def static(path):
-    global _static
-    if _static is None:
-        if apps.is_installed('django-contrib.staticfiles'):
-            from django.contrib.staticfiles.templatetags.staticfiles import static as _static
-        else:
-            from django.templatetags.static import static as _static
-    if django.VERSION >= (1, 9) and path == 'admin/img/icon-unknown.gif':
-        path = 'admin/img/icon-unknown.svg'
-    return _static(path)
 
 
 @register.filter